object_detection_transformer/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
Pillow>=9.5.0
pyyaml>=6.0
opencv-python>=4.9.0
streamlit>=1.30.0
pandas>=1.5.0
networkx>=3.0
